+++ packages/xlsx-preview/src/xlcore_wasm_bg.wasm.d.ts
@@ -7,343 +7,93 @@ export const extract_csv: (a: number, b: number, c: number, d: number) => void;
 export const extract_parquet: (a: number, b: number, c: number, d: number) => void;
 export const extract_xlsx: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_addMerge: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_addThreadedNote: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_addThreadedNote: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_autoFilter: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_calcProperties: (a: number, b: number) => void;
 export const workbookhandle_charts: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_clear: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_clearConditionalFormats: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-) => void;
+export const workbookhandle_clearConditionalFormats: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_clearRange: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_clearRangeWith: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_clearWith: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_clearRangeWith: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_clearWith: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_comments: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_conditionalFormats: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-) => void;
-export const workbookhandle_copyRange: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_conditionalFormats: (a: number, b: number, c: number, d: number) => void;
+export const workbookhandle_copyRange: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_createSheet: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_dataValidations: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_definedNames: (a: number, b: number) => void;
-export const workbookhandle_deleteColumns: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_deleteRows: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_deleteColumns: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_deleteRows: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_deleteSheet: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_dependencies: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_dependents: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_dispose: (a: number) => void;
-export const workbookhandle_fillRange: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_fillRange: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_getCell: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_getFreeze: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_getRange: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_hyperlinks: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_images: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_insertColumns: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_insertRows: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_insertColumns: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_insertRows: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_layout: (a: number, b: number, c: number) => void;
 export const workbookhandle_merges: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_moveSheet: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_moveSheet: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_new: (a: number) => void;
 export const workbookhandle_open: (a: number, b: number, c: number) => void;
 export const workbookhandle_pageSetup: (a: number, b: number, c: number, d: number) => void;
+export const workbookhandle_pivots: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_precedents: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_properties: (a: number, b: number) => void;
 export const workbookhandle_recalculate: (a: number, b: number) => void;
 export const workbookhandle_removeAutoFilter: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_removeAutoFilterColumn: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_removeChart: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_removeAutoFilterColumn: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_removeChart: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_removeComment: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_removeDataValidation: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-) => void;
-export const workbookhandle_removeDefinedName: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_removeDataValidation: (a: number, b: number, c: number, d: number) => void;
+export const workbookhandle_removeDefinedName: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_removeHyperlink: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_removeImage: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_removeImage: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_removeMerge: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_removePageSetup: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_removeSheetProtection: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-) => void;
-export const workbookhandle_removeSparklineGroup: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_removePivot: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_removeSheetProtection: (a: number, b: number, c: number, d: number) => void;
+export const workbookhandle_removeSparklineGroup: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_removeTable: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_removeThreadedThread: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-) => void;
+export const workbookhandle_removeThreadedThread: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_removeWorkbookProtection: (a: number, b: number) => void;
-export const workbookhandle_renameSheet: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_replyThreadedNote: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_renameSheet: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_replyThreadedNote: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_save: (a: number, b: number) => void;
 export const workbookhandle_search: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setActiveSheet: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_setAutoFilter: (a: number, b: number, c: number, d: number) => void;
-export const workbookhandle_setAutoFilterColumn: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setAutoFilterColumn: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setCalcProperties: (a: number, b: number, c: number) => void;
 export const workbookhandle_setChart: (a: number, b: number, c: number) => void;
-export const workbookhandle_setColumnVisible: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setColumnWidth: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setComment: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_setConditionalFormat: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_setDataValidation: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setColumnVisible: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setColumnWidth: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setComment: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setConditionalFormat: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setDataValidation: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setDefinedName: (a: number, b: number, c: number) => void;
-export const workbookhandle_setFormula: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setFreeze: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setHyperlink: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setFormula: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setFreeze: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setHyperlink: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setImage: (a: number, b: number, c: number) => void;
-export const workbookhandle_setPageSetup: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setPageSetup: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setPivot: (a: number, b: number, c: number) => void;
 export const workbookhandle_setProperties: (a: number, b: number, c: number) => void;
-export const workbookhandle_setRangeFormulas: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_setRangeValues: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_setRowHeight: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setRowVisible: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
-export const workbookhandle_setSheetProtection: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
-export const workbookhandle_setSheetVisibility: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-  f: number,
-) => void;
+export const workbookhandle_setRangeFormulas: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setRangeValues: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setRowHeight: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setRowVisible: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
+export const workbookhandle_setSheetProtection: (a: number, b: number, c: number, d: number, e: number) => void;
+export const workbookhandle_setSheetVisibility: (a: number, b: number, c: number, d: number, e: number, f: number) => void;
 export const workbookhandle_setSparklineGroup: (a: number, b: number, c: number) => void;
-export const workbookhandle_setStyle: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setStyle: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setTable: (a: number, b: number, c: number) => void;
-export const workbookhandle_setValue: (
-  a: number,
-  b: number,
-  c: number,
-  d: number,
-  e: number,
-) => void;
+export const workbookhandle_setValue: (a: number, b: number, c: number, d: number, e: number) => void;
 export const workbookhandle_setWorkbookProtection: (a: number, b: number, c: number) => void;
 export const workbookhandle_sheetProtection: (a: number, b: number, c: number, d: number) => void;
 export const workbookhandle_sheets: (a: number, b: number) => void;
@@ -356,5 +106,5 @@ export const workbookhandle_workbookProtection: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;
-export const __wbindgen_add_to_stack_pointer: (a: number) => number;
 export const __wbindgen_export4: (a: number, b: number, c: number) => void;
+export const __wbindgen_add_to_stack_pointer: (a: number) => number;